% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_date_suffix.R
\name{append_date_suffix}
\alias{append_date_suffix}
\title{append_date_suffix}
\usage{
append_date_suffix(dates)
}
\arguments{
\item{dates}{A vector of dates for which to generate the appropriate day of the month and assocaited suffix}
}
\value{
A single value that is the estimate mode.
}
\description{
Generates a day of the month value with the approriate suffix (e.g. 9th or 2nd)
Obtained from https://stackoverflow.com/questions/40039903/r-add-th-rd-and-nd-to-dates
}
