% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dredging_wq_transformations.R
\name{assd.xform}
\alias{assd.xform}
\title{assd.xform}
\usage{
assd.xform(x)
}
\arguments{
\item{x}{A numeric vector of raw ssd values}
}
\value{
A numeric vector of transformed values.
}
\description{
Applies a transformation
}
