% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chains.R
\name{check.chains}
\alias{check.chains}
\title{check.chains}
\usage{
check.chains(X, params, name = "")
}
\arguments{
\item{X}{a jag model fit as returned by a call to jags from R2jags}

\item{params}{A vector of names indicating the parameters that were traced during the jags fit}

\item{name}{and option character string containing any label to be placed at the top of the plot}
}
\value{
A plot of MCMC chains and ACF diagrames for each element in params.
}
\description{
Generates a plot of MCMC chains and ACF function for a Jags model output.
}
