% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tibble_colnames.R
\name{clean.tibble.colnames}
\alias{clean.tibble.colnames}
\title{clean.tibble.colnames}
\usage{
clean.tibble.colnames(x)
}
\arguments{
\item{x}{The tibble or data.frame to clean}
}
\value{
x with column names stripped of illegal characters.
}
\description{
Function to remove illegal characters in tibbles imported through tidyverse methods.
}
\details{
Removes illegal characters in column names allowed in tibbles but
would not normally be allowed in base R. this is to make them usable with code
that odes not support this functionality of tibbles. Currently replaces " " with "."
and all other characters with "".
}
