% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dredging_wq_transformations.R
\name{dli.xform}
\alias{dli.xform}
\title{dli.xform}
\usage{
dli.xform(x)
}
\arguments{
\item{x}{A numeric vector of raw DLI values}
}
\value{
A numeric vector of transformed values.
}
\description{
Applies a transformation
}
