% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSassymp_select.R
\name{fit.SSasymp}
\alias{fit.SSasymp}
\title{fit.SSasymp}
\usage{
fit.SSasymp(data, exp.val = 2)
}
\arguments{
\item{data}{The data frame containing the x and y to which the model should be fit.
Note these must currently by labelled x and y}

\item{exp.val}{The exponent to use for generate the search grid of potential
starting values. Large values will make the function very slow to run.}
}
\value{
A list of fitted assymptotic models that can be sued for multimodel inference or
to select the best model (ie based on the lowest AICc).
}
\description{
Fits a range of assymptotic models and returns a list of model fits.
}
