% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse_logit.R
\name{inverse.logit}
\alias{inverse.logit}
\title{inverse.logit}
\usage{
inverse.logit(x, car.rescale = F)
}
\arguments{
\item{x}{A vector of values for which to apply a logit back-transformation}

\item{car.rescale}{Ture/FALSE indicator for if car based rescaling is required.}
}
\value{
A vector of logit-backtransformed values.
}
\description{
Backtransforms a logit transformation that incolved rescaling as per car.
}
