% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dredging_wq_transformations.R
\name{ntu.xform}
\alias{ntu.xform}
\title{ntu.xform}
\usage{
ntu.xform(x)
}
\arguments{
\item{x}{A numeric vector of raw NTU values}
}
\value{
A numeric vector of transformed values.
}
\description{
Applies a transformation
}
