% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FSS_mod.R
\name{plot.FSS.mod}
\alias{plot.FSS.mod}
\title{plot.FSS.mod}
\usage{
\method{plot}{FSS.mod}(mod.r, y.lab = "response")
}
\arguments{
\item{mod.r}{The gam model objsect to plot}

\item{y.lab}{An optional Y label for the response variable}
}
\value{

}
\description{
Generates a plot on the raw scale of the response, for a gam model
}
