% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_plot_mat.R
\name{poly.plot.mat}
\alias{poly.plot.mat}
\title{poly.plot.mat}
\usage{
poly.plot.mat(x, start.col, end.col, y.lim, alpha.f = 1,
  col.val = "yellow")
}
\arguments{
\item{x}{The matrix/data.frame from which to obtain values for plotting}

\item{start.col}{The column in the matrix or data.frame that defines the start plotting value for each row}

\item{end.col}{The column in the matrix or data.frame that defines the end plotting value for each row}

\item{y.lim}{The y values over which to draw the polygons}
}
\description{
Plot polygons on the existing plot from a supplied matrix of values
}
