% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nan.R
\name{remove_nan}
\alias{remove_nan}
\title{remove_nan}
\usage{
remove_nan(x)
}
\arguments{
\item{x}{A Numeric vector}
}
\value{
A numeric vector with any NaN values replaced with NA
}
\description{
Removes NaN and replaces with NA in a numeric vector
}
