% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wi.R
\name{wi}
\alias{wi}
\title{wi}
\usage{
wi(AIC.vals)
}
\arguments{
\item{AIC.vals}{A vector of AIC, AICc or BIC values}
}
\value{
A vector of Akaike weights
}
\description{
Calculates model weights based on AIC, AICc, or BIC
}
