% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dredging_wq_transformations.R
\name{xform}
\alias{xform}
\title{xform}
\usage{
xform(x, var.name)
}
\arguments{
\item{x}{A numeric vector of raw NTU values}

\item{var.name}{A character string indicating the variable being transformed}
}
\value{
A numeric vector of transformed values.
}
\description{
Applies a transformation
}
