% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xformEta.R
\name{xformEta}
\alias{xformEta}
\title{xformEta}
\usage{
xformEta(x, variableType, direction)
}
\arguments{
\item{x}{A vector of values to transform}

\item{variableType}{A character string indicating the type of distribution that has been used. Must be one of
'gaussian', 'poisson', 'gamma', 'nbinomial', 'binomial', or 'beta'.}

\item{direction}{A character string indication the direction to tranform x. Must be either 'response', or 'link'.
For 'response' x is assumed to be on the scale of the link function, and is backtransformed to the response scale.
For 'link' x is assumed to be on the scale of the response, and is transformed to the link scale.
Assumed link fuctions are as follows: 'gaussian' = identity, 'poisson' = log, 'gamma'  = log, 'nbinomial' - log, 'binomial' = logit, 'beta' = logit.}
}
\value{
A vector of Akaike weights
}
\description{
Transforms and backtransforms between commonly link functions for a range of statistical distributions
}
