% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc.mode}
\alias{calc.mode}
\title{calc.mode}
\usage{
calc.mode(x, type = 1)
}
\arguments{
\item{x}{A vector of values for which to calculate a mode}

\item{type}{The type of mode to calculate for numeric vectors}
}
\value{
A single value that is the estimate mode.
}
\description{
Calculates a model value, using the most appropriate method given the data type.
}
